--- backend/typescript/middlewares/auth.graphql.ts
@@ -1,100 +1,0 @@
-import { AuthenticationError, ExpressContext } from "apollo-server-express";
-import { GraphQLResolveInfo } from "graphql";
-
-import { getAccessToken } from "./auth";
-import AuthService from "../services/implementations/authService";
-import UserService from "../services/implementations/userService";
-import IAuthService from "../services/interfaces/authService";
-import { Role } from "../types";
-
-const authService: IAuthService = new AuthService(new UserService());
-
-/* Determine if request is authorized based on accessToken validity and role of client */
-export const isAuthorizedByRole = (roles: Set<Role>) => {
-  return async (
-    resolve: (
-      parent: any,
-      args: { [key: string]: any },
-      context: ExpressContext,
-      info: GraphQLResolveInfo,
-    ) => any,
-    parent: any,
-    args: { [key: string]: any },
-    context: ExpressContext,
-    info: GraphQLResolveInfo,
-  ) => {
-    const accessToken = getAccessToken(context.req);
-    const authorized =
-      accessToken && (await authService.isAuthorizedByRole(accessToken, roles));
-
-    if (!authorized) {
-      throw new AuthenticationError(
-        "Failed authentication and/or authorization by role",
-      );
-    }
-
-    return resolve(parent, args, context, info);
-  };
-};
-
-/* Determine if request for a user-specific resource is authorized based on accessToken
- * validity and if the userId that the token was issued to matches the requested userId
- * Note: userIdField is the name of the request parameter containing the requested userId */
-export const isAuthorizedByUserId = (userIdField: string) => {
-  return async (
-    resolve: (
-      parent: any,
-      args: { [key: string]: any },
-      context: ExpressContext,
-      info: GraphQLResolveInfo,
-    ) => any,
-    parent: any,
-    args: { [key: string]: any },
-    context: ExpressContext,
-    info: GraphQLResolveInfo,
-  ) => {
-    const accessToken = getAccessToken(context.req);
-    const authorized =
-      accessToken &&
-      (await authService.isAuthorizedByUserId(accessToken, args[userIdField]));
-
-    if (!authorized) {
-      throw new AuthenticationError(
-        "Failed authentication and/or authorization by userId",
-      );
-    }
-
-    return resolve(parent, args, context, info);
-  };
-};
-
-/* Determine if request for a user-specific resource is authorized based on accessToken
- * validity and if the email that the token was issued to matches the requested email
- * Note: emailField is the name of the request parameter containing the requested email */
-export const isAuthorizedByEmail = (emailField: string) => {
-  return async (
-    resolve: (
-      parent: any,
-      args: { [key: string]: any },
-      context: ExpressContext,
-      info: GraphQLResolveInfo,
-    ) => any,
-    parent: any,
-    args: { [key: string]: any },
-    context: ExpressContext,
-    info: GraphQLResolveInfo,
-  ) => {
-    const accessToken = getAccessToken(context.req);
-    const authorized =
-      accessToken &&
-      (await authService.isAuthorizedByEmail(accessToken, args[emailField]));
-
-    if (!authorized) {
-      throw new AuthenticationError(
-        "Failed authentication and/or authorization by email",
-      );
-    }
-
-    return resolve(parent, args, context, info);
-  };
-};
