+++ backend/typescript/models/entity.ts
@@ -1,3 +1,4 @@
+// mongodb {
 import { Schema, Document, model } from "mongoose";
 
 export interface Entity extends Document {
@@ -34,3 +35,29 @@ const EntitySchema: Schema = new Schema({
 });
 
 export default model<Entity>("Entity", EntitySchema);
+
+// } mongodb
+// postgresql {
+import { Column, Model, Table, DataType } from "sequelize-typescript";
+
+import { Letters } from "../types";
+
+@Table({ tableName: "entities" })
+export default class Entity extends Model {
+  @Column
+  string_field!: string;
+
+  @Column
+  int_field!: number;
+
+  @Column({ type: DataType.ENUM("A", "B", "C", "D") })
+  enum_field!: Letters;
+
+  @Column({ type: DataType.ARRAY(DataType.STRING) })
+  string_array_field!: string[];
+
+  @Column
+  bool_field!: boolean;
+}
+
+// } postgresql
