+++ backend/typescript/models/user.ts
@@ -1,3 +1,4 @@
+// mongodb {
 import mongoose, { Schema, Document } from "mongoose";
 
 import { Role } from "../types";
@@ -31,3 +32,25 @@ const UserSchema: Schema = new Schema({
 });
 
 export default mongoose.model<User>("User", UserSchema);
+
+// } mongodb
+// postgresql {
+import { Column, DataType, Model, Table } from "sequelize-typescript";
+import { Role } from "../types";
+
+@Table({ tableName: "users" })
+export default class User extends Model {
+  @Column({ type: DataType.STRING })
+  first_name!: string;
+
+  @Column({ type: DataType.STRING })
+  last_name!: string;
+
+  @Column({ type: DataType.STRING })
+  auth_id!: string;
+
+  @Column({ type: DataType.ENUM("User", "Admin") })
+  role!: Role;
+}
+
+// } postgresql
